+++ src/Module/WPLoader/FiltersGroup.php
@@ -5,7 +5,7 @@
  * @package tad\WPBrowser\Module\WPLoader
  */
 
-namespace tad\WPBrowser\Module\WPLoader;
+namespace lucatume\WPBrowser\Module\WPLoader;
 
 /**
  * Class FiltersGroup
@@ -14,13 +14,6 @@ namespace tad\WPBrowser\Module\WPLoader;
  */
 class FiltersGroup
 {
-    /**
-     * An array detailing each filter callback, priority and arguments.
-     *
-     * @var array<array<mixed>>
-     */
-    protected $filters = [];
-
     /**
      * The callback that will be used to remove filters.
      *
@@ -43,19 +36,19 @@ class FiltersGroup
      *                                        the default one.
      * @param callable|null       $addWith    The callable that should be used to add the filters, or `null` to use the
      */
-    public function __construct(array $filters = [], callable $removeWith = null, callable $addWith = null)
+    public function __construct(/**
+     * An array detailing each filter callback, priority and arguments.
+     */
+    protected array $filters = [], callable $removeWith = null, callable $addWith = null)
     {
-        $this->filters        = $filters;
-        $this->removeCallback = null === $removeWith ? 'remove_filter' : $removeWith;
-        $this->addCallback    = null === $addWith ? 'add_filter' : $addWith;
+        $this->removeCallback = $removeWith ?? 'remove_filter';
+        $this->addCallback    = $addWith ?? 'add_filter';
     }
 
     /**
      * Removes the filters of the group.
-     *
-     * @return void
      */
-    public function remove()
+    public function remove(): void
     {
         foreach ($this->filters as $filter) {
             $filterWithoutAcceptedArguments = array_slice($filter, 0, 3);
@@ -65,10 +58,8 @@ class FiltersGroup
 
     /**
      * Adds the filters of the group.
-     *
-     * @return void
      */
-    public function add()
+    public function add(): void
     {
         foreach ($this->filters as $filter) {
             call_user_func_array($this->addCallback, $filter);