+++ tests/dbunit/WPHealthcheckTest.php
@@ -2,12 +2,15 @@
 
 namespace tad\WPBrowser\Module\Support;
 
+use lucatume\WPBrowser\Module\Support\WordPressDatabase;
+use lucatume\WPBrowser\Module\Support\WordPressDirectories;
+use lucatume\WPBrowser\Module\Support\WPHealthcheck;
 use tad\Codeception\SnapshotAssertions\SnapshotAssertions;
 use tad\Test\Constants as TestConstants;
 use tad\WPBrowser\Environment\Constants;
 use tad\WPBrowser\Generators\Tables;
-use tad\WPBrowser\Traits\WithStubProphecy;
 use tad\WPBrowser\StubProphecy\Arg;
+use tad\WPBrowser\Traits\WithStubProphecy;
 use function tad\WPBrowser\envFile;
 use function tad\WPBrowser\Tests\Support\testEnvFile;