+++ tests/unit/tad/WPBrowser/Module/Connector/WordPressTest.php
@@ -1,12 +1,12 @@
 <?php
 namespace tad\WPBrowser\Module\Connector;
 
+use lucatume\WPBrowser\Module\Support\UriToIndexMapper;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
 use Symfony\Component\BrowserKit\CookieJar;
 use Symfony\Component\BrowserKit\History;
 use tad\WPBrowser\Connector\WordPress;
-use tad\WPBrowser\Module\Support\UriToIndexMapper;
 use tad\WPBrowser\Traits\WithStubProphecy;
 
 class WordPressTest extends \Codeception\Test\Unit
@@ -34,7 +34,7 @@ class WordPressTest extends \Codeception\Test\Unit
     protected $cookieJar;
 
     /**
-     * @var UriToIndexMapper
+     * @var \lucatume\WPBrowser\Module\Support\UriToIndexMapper
      */
     protected $uriToIndexMapper;